requirements.txt
python-dotenv>=0.12.0,<0.13.0
spotipy>=2.10.0,<2.11.0
Flask>=1.1.1,<1.2.0
# opencv-python==4.2.0.32
Flask-SQLAlchemy>=2.4.1,<2.5.0
Flask-Session>=0.3.1,<0.4.0
psycopg2>=2.8.4,<2.9.0